pyproject.toml
[tool.poetry]
name = "airsenal"
version = "1.10.0"
description = "AI manager for Fantasy Premier League"
authors = [
	"Angus Williams <anguswilliams91@gmail.com>",
	"Jack Roberts <jroberts@turing.ac.uk>",
	"Nick Barlow <nbarlow@turing.ac.uk>",
]

[tool.poetry.dependencies]
python = "^3.9"
bpl = { git = "https://github.com/anguswilliams91/bpl-next" }
click = "^8.1.7"
pandas = "^2.2.1"
requests = "^2.31.0"
sqlalchemy = "^2.0.27"
tqdm = "^4.66.2"
dateparser = "^1.2.0"
prettytable = "^3.10.0"
beautifulsoup4 = "^4.12.3"
platformdirs = "^4.2.0"
jupyterlab = {version = "^4.1.2", optional = true}
matplotlib = {version = "^3.8.3", optional = true}
Flask = {version = "^3.0.2", optional = true}
numpyro = {version = "^0.13.2", optional = true}
seaborn = {version = "^0.13.2", optional = true}
deap = {version = "^1.4.1", optional = true}
python-dateutil = "^2.8.2"
lxml = "^5.1.0"
html5lib = "^1.1"
jax = "^0.4.24"
jaxlib = "^0.4.24"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
pre-commit = "^3.6.2"
pytest = "^8.0.1"
pytest-sugar = "^1.0.0"

[tool.poetry.extras]
api = ["Flask"]
notebook = ["jupyterlab"]
plot = ["matplotlib", "seaborn"]
optimization = ["deap"]

[build-system]
requires = ["poetry-core>=1.0.0", "cython", "numpy", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = """
--cov=airsenal
--cov-report=term:skip-covered
--pdbcls=IPython.terminal.debugger:TerminalPdb
"""

[tool.isort]
profile = "black"

[tool.poetry.scripts]
airsenal_run_pipeline = "airsenal.scripts.airsenal_run_pipeline:run_pipeline"
airsenal_setup_initial_db = "airsenal.scripts.fill_db_init:main"
airsenal_update_db = "airsenal.scripts.update_db:main"
airsenal_run_prediction = "airsenal.scripts.fill_predictedscore_table:main"
airsenal_make_squad = "airsenal.scripts.squad_builder:main"
airsenal_run_optimization = "airsenal.scripts.fill_transfersuggestion_table:main"
airsenal_make_transfers = "airsenal.scripts.make_transfers:main"
airsenal_set_lineup = "airsenal.scripts.set_lineup:main"
airsenal_replay_season = "airsenal.scripts.replay_season:main"
airsenal_env = "airsenal.scripts.set_env:main"
airsenal_check_data = "airsenal.scripts.data_sanity_checks:run_all_checks"
airsenal_dump_api = "airsenal.scripts.dump_api:main"
airsenal_dump_db = "airsenal.scripts.dump_db_contents:main"
airsenal_scrape_transfermarkt = "airsenal.scripts.scrape_transfermarkt:main"
airsenal_plot = "airsenal.scripts.plot_league_standings:main"
